tcpc shim: Verify the TCPC compatible is valid
The macro that creates the tcpc_config[] array builds without an error
even when the "tcpc" property is present, but the reference node
contains an invalid compatible string.  This resulted in a NULL entry in
the tcpc_config() array and causes crashes or boot loops.

Add a check that generates a build error if the "tcpc" property is
present, but fails to resolve to a supported TCPC driver.

Devicetree snippet:

	tcpc_port0: rt1715@4e {
		compatible = "richtek,rt1715-tcpc-invalid";
		reg = <0x4e>;
		tcpc-flags = <(
			TCPC_FLAGS_TCPCI_REV2_0 |
			TCPC_FLAGS_TCPCI_REV2_0_NO_VSAFE0V)>;
		irq-gpios = <&gpio3 7 GPIO_ACTIVE_LOW>;
	};

Generates the following error:
  error: static assertion failed:
    "Invalid TCPC compatible on node:
    DT_N_S_soc_if_S_io_i2c_ctrl0_port0_S_rt1715_4e"

BUG=b:283134516
TEST=zmake build -a

Change-Id: I09257a40ba84e1fdda078faa9fea608f6da51124
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4550244
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>