test/cortexm_fpu: Fix underflow test when LTO is enabled
The interrupt for the FPU is asynchronous. When building with LTO, there
are fewer instructions between triggering the interrupt with software on
H7 (task_trigger_irq) and checking the fpscr variable, so it was failing
because the interrupt hadn't yet been handled to set the fpscr variable.

The only test that was failing was test_cortexm_fpu_underflow, and it
was only failing on icetower, but this change waits for the interrupt on
all the tests since it could theoretically happen on the others too.

BRANCH=none
BUG=b:274162810
TEST=./test/run_device_tests.py --board dartmonkey -t cortexm_fpu => PASS
TEST=./test/run_device_tests.py --board bloonchipper -t cortexm_fpu => PASS

Change-Id: I93e1f9a0166bcb271b68d21673b796089daf4fbe
Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4550028
Reviewed-by: Patryk Duda <patrykd@google.com>