zephyr: gpio: Don't create gpio_dt_spec pointer for unimplemented GPIO
If the GPIO is unimplemented, then GPIO_SIGNAL(id) is resolved to an
gpio_signal enum entry which has a value of -1 (GPIO_UNIMPLEMENTED).
As a result we point to address outside of the configs array. Also some
compilers don't like negative index.

This patch adds check if GPIO DT node has 'gpios' property in
GPIO_PTRS() and GPIO_DT_PTR_DECL() macros. The pointer will be created
only if the condition is satisfied.

BUG=b:239712345
TEST=zmake build bloonchipper (with 'llvm' as toolchain)

Change-Id: Ia66c90c14ab0621bbf567b545362ec29a053a82b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4565516
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Patryk Duda <patrykd@google.com>