usb_common: Refuse PR swap when powered from Type-C and no battery
The dibbi pd_check_power_swap() function was copied from fizz, and
unlike the common function it doesn't allow PR swap when dual role is
disabled but we are sourcing, so SRC->SNK PR swap doesn't currently work
in S0ix.

The other difference is the fizz function also has a check to refuse PR
swap if the Type-C port is our active charge port. This is the behaviour
we want on all devices without batteries, so move this check to the
common function and change fizz and dibbi variants to use the common
function.

BUG=b:284926643
TEST=On dibbi:
- When powered from barrel jack, check that a SRC->SNK PR swap now works
  in S0ix.
- When powered from Type-C only, request a PR swap from the servo_v4
  console and check the EC refuses it.

Change-Id: If2d603f327911ef2e57f0808834fdfcfc3860610
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4576649
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Tested-by: Reka Norman <rekanorman@chromium.org>