charger: Move params for THROTTLE_AP into the charger
Only Scarlet defines these so there is no need for them to be
configurable. New boards use Zephyr which has this value in a
Kconfig option, but again this not used. It may never be used.

Move the (only) values into the charger source instead. Drop the
Kconfig options.

BUG=b:218332694
TEST=make buildall

Change-Id: I7b1900c6c3183a453bcd7ba23a3d83f01f67a34d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4577900
Commit-Queue: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Aaron Massey <aaronmassey@google.com>