zephyr: rex: kbd: set up the pins for the kbd node
This set up the dts node for the upstream keyboard driver in addition to
the current one in the shim. The upstream driver is not used in the EC
at this stage so this is a no-op, but it's one less change to start
experimenting with it.

BUG=b:274443029
TEST=zmake compare-builds -a

Change-Id: I2f4944124063dae01d2cef73440b9d10098c9eeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4580222
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>