zephyr/motionsense: ensure default_range is signed
Not casting the value of default_range from the device tree to a signed
value causes clang to raise a warning (promoted to an error) regarding
narrowing conversions if using ROUND_UP_FLAG. ROUND_UP_FLAG is defined
to be an unsigned value, so ROUND_UP_FLAG | <literal integer> is
interpreted as unsigned which then causes a warning because
the resulting value is converted from a large positive number to a
large negative one.

It would generally be more correct to make the range unsigned, but it's
difficult to ensure that correct behavior is retained because some code
paths do arithmetic on signed sensor readings and changing the range
fields of struct motion_sensor_t can have surprising effects on those
operations. Multiplying or dividing a signed number (like a sensor
reading) by an unsigned one (the range) causes the signed value to be
promoted to an unsigned one first which will lead to anomalously large
computed values anytime the input value is negative.

BUG=None
TEST=twister

Change-Id: I3ac6755a1e76321785893f015dbd3dd58db53af6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4583469
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Peter Marheine <pmarheine@chromium.org>
Commit-Queue: Peter Marheine <pmarheine@chromium.org>