Skyrim: Enable CONFIG_PLATFORM_EC_EMULATED_SYSRQ
Enable CONFIG_PLATFORM_EC_EMULATED_SYSRQ for Skyrim, to allow the EC to
send SysRq signals to the AP. When the AP receives a SysRq signal, it
will generate a stack trace for each of the CPU cores, which is useful
when attempting to debug AP hangs.

The EC suspend/resume hang detection can now be updated to send a SysRq,
so we can (hopefully) have some AP data to debug the issue with.

BUG=b:284546452
TEST=ec:~$ sysrq

Change-Id: If0b42af130e1d0a1f722995bbaf04660acb736ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4591415
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Tim Van Patten <timvp@google.com>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Tim Van Patten <timvp@google.com>