npcx: cec: Restructure code before moving to common
Restructure the code in preparation for moving it to common: add a cec
prefix to some functions/types, and pull the chip-specific timer
functionality out into separate functions. This allows the changes in
the following CL to be pure copy-pastes for ease of review. No
functional changes intended.

BUG=b:270507438
TEST=Turn a TV off and on using CEC on kinox

Change-Id: I10b347c58a5911d3a88d1a99456076caec3a9456
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4600563
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Tested-by: Reka Norman <rekanorman@chromium.org>