cec: Restructure the bit-banging CEC implementation as a driver
Add a cec_drv type and restructure the bit-banging CEC implementation as
an instance of this driver. This will allow other CEC drivers to be
added, specifically a driver for the CEC peripheral on the ITE EC.

For now the driver is in common/cec.c. The following CL will move it to
a separate file. This is done in two steps just for ease of review.

Allow boards to configure the number of CEC ports and the driver used
for each.

For now the common code (task, host command handlers, etc) only supports
one port, so all operations are performed on port 0. Support for
multiple ports will be added in a later CL.

BUG=b:270507438
TEST=Use CEC to turn a TV on and off on dibbi and kinox.
LOW_COVERAGE_REASON=Tests will be added in b:287558802

Change-Id: I7dc239bf2085bdafd783229eaa7c69c4eed7a117
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4626200
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Reka Norman <rekanorman@chromium.org>
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>