third_party/boringssl: Fix linker warning
Make sure that boringssl is compiled with the same flags as the rest of
the EC code. Otherwise, the linker warns that we're linking modules of
different target triples: 'thumbv7m-cros-unknown-eabi' and
'thumbv7em-cros-unknown-eabihf'.

This change will not be needed once https://crrev.com/c/4610318 lands,
but that is blocked on upstream boringssl changes.

BUG=b:287672777, b:287661706
TEST=make BOARD=bloonchipper
TEST=make BOARD=dartmonkey

Change-Id: Iae7cd20fa519ce47c868e061fccc9fa12029ec59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4638137
Reviewed-by: Craig Hesling <hesling@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>