core/cortex-m[0]: Add volatile to inline asm in interrupt functions
When enabling LTO for bloonchipper/dartmonkey, we found that a call to
is_interrupt_enabled() was optimized out, causing failures.

Examining the code, we see that the problem is that the compiler assumes
that it can cache the result from the call to is_interrupt_enabled()
because it doesn't know that the inline assembly modifies the special
"primask" register. We could solve this by adding "primask" to the
clobber list, but the "primask" register is not one that clang is aware
of. Adding "volatile" on reads tells the compiler that it must perform
as many reads as the program specifies (e.g., it can't reuse the result
of an old one).

Examining other code in this file, we see that interrupt_enable() and
interrupt_disable() also should have "volatile" since they read/write
"primask". For writes, "volatile" tells the compiler that it can't
re-order the writes relative to other volatile operations (such as the
reads in the other functions).

The case of in_interrupt_context() is similar to is_interrupt_enabled(),
with the addition that "ipsr" can change asynchronously (exceptions,
etc.), so we need to make sure that the compiler performs a read each
time.

BUG=b:287532311
TEST=./util/compare_build.sh -b dartmonkey => MATCH
TEST=./test/run_device_tests.py --board bloonchipper
     Test "production_app_test": PASSED
     Test "abort": PASSED
     Test "aes": PASSED
     Test "always_memset": PASSED
     Test "benchmark": PASSED
     Test "boringssl_crypto": PASSED
     Test "cortexm_fpu": PASSED
     Test "crc": PASSED
     Test "exception": PASSED
     Test "flash_physical": PASSED
     Test "flash_write_protect": PASSED
     Test "fpsensor_auth_crypto_stateless": PASSED
     Test "fpsensor_hw": PASSED
     Test "fpsensor_spi_ro": PASSED
     Test "fpsensor_spi_rw": PASSED
     Test "fpsensor_uart_ro": PASSED
     Test "fpsensor_uart_rw": PASSED
     Test "ftrapv": PASSED
     Test "libc_printf": PASSED
     Test "global_initialization": PASSED
     Test "libcxx": PASSED
     Test "malloc": PASSED
     Test "mpu_ro": PASSED
     Test "mpu_rw": PASSED
     Test "mutex": PASSED
     Test "panic": PASSED
     Test "pingpong": PASSED
     Test "printf": PASSED
     Test "queue": PASSED
     Test "rng_benchmark": PASSED
     Test "rollback_region0": PASSED
     Test "rollback_region1": PASSED
     Test "rollback_entropy": PASSED
     Test "rtc": PASSED
     Test "sbrk": PASSED
     Test "sha256": PASSED
     Test "sha256_unrolled": PASSED
     Test "static_if": PASSED
     Test "stdlib": PASSED
     Test "std_vector": PASSED
     Test "stm32f_rtc": PASSED
     Test "system_is_locked_wp_on": PASSED
     Test "system_is_locked_wp_off": PASSED
     Test "timer": PASSED
     Test "timer_dos": PASSED
     Test "tpm_seed_clear": PASSED
     Test "utils": PASSED
     Test "utils_str": PASSED
     Test "panic_data_bloonchipper_v2.0.4277": PASSED
     Test "panic_data_bloonchipper_v2.0.5938": PASSED

Change-Id: Iaeedabe793e963c1959520dc370da904b0c21a02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4652124
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Peter Marheine <pmarheine@chromium.org>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>
Tested-by: Tom Hughes <tomhughes@chromium.org>