do not set ext power enabled when no battery
pd controller seems to have modified behavior from ccg6 so that
when a port is configured as a sink and extpower is present
it will exit dead battery mode.
Causing system cannot power on in standalone mode.
Do not set extpower present on controller that is sinking power
in standalone mode

Signed-off-by: Kieran Levin <ktl@frame.work>