bazel: Move twister bazel here
It's easier on people if we can put the build rules for this repo in
the repo itself.  This migrates the twister build rules into this
repository in the Bazel directory.

The copy of this file in rules_cros_firmware will be deleted after
this lands.

BUG=b:274648575
TEST=./twister -iv -T zephyr/test/hooks

Change-Id: I47fe34e2d114a6a371b0a3332f10097323b020a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4669751
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>