bazel: Move zmake bazel here
It's easier on people if we can put the build rules for this repo in
the repo itself.  This migrates the zmake build rules into this
repository in the Bazel directory.

The copy of this file in rules_cros_firmware will be deleted after
this lands.

BUG=b:274648575
TEST=bazel build :brya

Change-Id: Iaa1615d47f24d4b6c071679aabba6797c1ed97cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4670199
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>