twister_launcher: Couple TOOLCHAIN cmake/env vars
The TOOLCHAIN_HOME variable as a cmake variable and as an environment
variable are both coupled as they represent the same thing. They only
exist separately as cmake and environment variables due to
verify-toolchain.cmake requiring it as an environment variable. At
present in the twister_launcher script, the environment variable and the
cmake variable are not coupled as they ought to be.

Define TOOLCHAIN_ROOT and use the same value as an environment and cmake
variable when calling twister.

BUG=b:268051194,b:268051194
TEST=twister -T zephyr/test/uart_printf/ --coverage #unittest
TEST=twister -T zephyr/test/system_shim/ --coverage

Change-Id: Ifbe3f031f4ab9fb1399eb0b37091b42433da11d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4679624
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>