zephyr: tests: do not rely on the response buffer address
Change tests to rely on the content of the host command response buffer
instead of an address of the buffer.

The change is needed to use the upstream host command system, which uses
an internal buffers for request and response.

BUG=b:286966639
TEST=./twister -v -i  -p native_posix -p unit_testing

Change-Id: I89cbf28d22683d76efadd27fcc320e93a5940eef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4737957
Tested-by: Dawid Niedzwiecki <dawidn@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Dawid Niedzwiecki <dawidn@google.com>