common: Add AMD STB dump feature flag
AMD SoCs have a low-level log called the STB. If supported, the EC can
trigger the SoC to save the contents of the STB to help with debugging.
Add a feature flag for this functionality to aid with testing.

BUG=b:258670202
TEST=Builds

Change-Id: I83d42c09fcbfc59ded1900c79dd96ff22853e417
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4826554
Tested-by: Robert Zieba <robertzieba@google.com>
Commit-Queue: Robert Zieba <robertzieba@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Forest Mittelberg <bmbm@google.com>