dibbi: Disable PPC when charging from BJ
Fix a copy-paste error in CL:4227433 so that the PPC is disabled when
charging from BJ (not enabled).

Also swap the order of operations when switching ports so that the
current port is disabled before enabling the new port. This reduces the
risk of voltage spikes during the switch.

BUG=b:291823427
TEST=When AP is in G3:
- Type-C connected, connect BJ -> switches to BJ successfully.
- Both connected, disconnect BJ -> EC browns out (previously it would
  stay powered and switch to Type-C).
Measurements of the hold up time, etc will be done later.

Change-Id: I3877b39d4fa98e362483bd5d7c62ba28d5c4b415
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4796749
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Reka Norman <rekanorman@chromium.org>