board/hyperdebug: Allow runtime reconfiguration of chip select
HyperDebug is a versatile device, operating in a similar way to uServo.
It makes several SPI ports abaiable for the host computer to control.

This CL adds a new console command for changing which pin is to be used
as the chip select signal, allowing HyperDebug to control multiple SPI
devices sitting on the same bus (that is, sharing data and clock
signals).

BUG=none
TEST=logic analyzer of manual runs with modified opentitantool

Change-Id: I7227687dd4d9d756b833bda7a2c26fb3454350ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4739792
Tested-by: Jes Klinke <jbk@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Jes Klinke <jbk@chromium.org>