nissa: test board_power.c
This needs to include ap_pwrseq machinery in order to define the power
signals as referred to by this code, but otherwise avoids actually
running the power sequencing state machine by defining every power signal
as external (generated by board code) which is then actually generated
by a test-controlled fake.

Changes will be required for AP_PWRSEQ_DRIVER support, but that's
not currently available for the board code so the test will not be
concerned with it either.

BUG=b:271118604
TEST=twister

Change-Id: I22aa80a2761d7ea5a33b306fa9e825f3f1dadc22
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4800415
Commit-Queue: Peter Marheine <pmarheine@chromium.org>
Tested-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>