rex: workaround the wrong device ID on certain boards
Some boards of the rex family use the sample batch of chips with the
wrong device ID. It casues the console command version print the
the "Unknown-092b" device information. This CL workarounds it by adding
a new config option to the problemaic board. When the option is
enabled, the system code will do the extra checking for that ID.

BUG=b:296945649
BRANCH=none
TEST="version" command can show the correct information.
TEST= pass "zmake build -a --clobber"
LOW_COVERAGE_REASON=this is the hardware workaround and there are no
tests for EC on-chip system driver.

Change-Id: I38420a6d719806de75fde172f9f03de69aa0c5e1
Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4805684
Reviewed-by: caveh jalali <caveh@chromium.org>
Commit-Queue: caveh jalali <caveh@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: CH Lin <chlin56@nuvoton.com>