extra: Remove special case for vidpid strings in servo common code
get_usb_dev() accepts a string (old style interface) or an iterable
object holding strings (new style) and switches based on the argument
type which is a code smell.

While this made sense during the transition so the logic could be
evaluated on its merits without dealing with the noise of adapting
all the call sites, get rid of this type dependent behavior now.

While at it, add type hints so that wrong calls in unexpected places
have a chance to be found (also for the usual reasons to add type
hints: discoverability of the API, general robustness)

BUG=b:170507074
TEST=searched through platform/ec and third_party/hdctools for users
of this API and adapted them. Also ran some manual tests with servo
4.1 where it identified the device, did not identify the device
(when asking for something else), etc.

Change-Id: I957b4a84cdc7261acc921600497b01e63098dc62
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4817009
Commit-Queue: Patrick Georgi <pgeorgi@google.com>
Reviewed-by: Matthew Blecker <matthewb@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Patrick Georgi <pgeorgi@google.com>