board/hyperdebug: Blue button resets shield
Neither the existing Dauntless shield, nor the future OpenTitan shield
will have a button to reset the security chip.  Instead, with this
change, HyperDebug will reset the security chip when the blue button is
pressed.

The feature is configurable, such that the user can choose any of
HyperDebug's pins to be pulled low, when the blue user button is
pressed.  The default, CN10_29, is the one that both Dauntless and
OpenTitan shields use for the RESET signal.

BUG=none
TEST=observe reset on Dauntless shield

Change-Id: I26237b92eb5d9a033c07b2201f5d733b055b95f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4833742
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Commit-Queue: Jes Klinke <jbk@chromium.org>
Tested-by: Jes Klinke <jbk@chromium.org>