board/hyperdebug: Enable SPI1
HyperDebug allows a host computer to control several SPI and I2C buses,
as well as UARTs and GPIOs all via a USB connection.  This CL adds the
SPI1 controller to the set that can be controller.

Unlike the other busses, the pins assigned for SPI1 will continue to act
as GPIO input on HyperDebug power-up, in order to not break existing
users of these pins.  Prior to issuing any SPI transactions on the SPI1
bus, the host should have executed commands such as "gpiosetmode CN7_15
alternate" on each of the pins to be used by SPI1 (and "gpiosetmode
... pushpull" on the chip select, possibly in conjunction with "spi set
cs SPI1 ...").

In practice, this will most conveniently be done by mentioning the pins
in a JSON configuration file for opentitantool.

BUG=b:299351632
TEST=manual inspection with logic analyzer

Change-Id: I5666db479194566f4b9953cfbd3bd27417abeb77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4852479
Reviewed-by: Jett Rink <jettrink@chromium.org>
Tested-by: Jes Klinke <jbk@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Jes Klinke <jbk@chromium.org>