board/hyperdebug: Encapsulate I2C into CMSIS-DAP
This change replaces the instantiation of a uServo-compatible I2C
tunneling USB interface with a board-specific implementation of the
CMSIS-DAP protocol.

The CMSIS-DAP is meant for controlling JTAG/SWD debugger devices, but
has provisions for "vendor extensions", the I2C tunnelling is now one
such extension.

The reason for doing this is that we want to add JTAG support to
HyperDebug, but we are at the limit of the number of USB interfaces that the
STM32L5 family of chips can support.

Besides the I2C interface, the other USB interfaces are used for UART
and for SPI forwarding.  The UART interfaces are declared such that they
show up as /dev/ttyUSBn, and we do not want to change that.  The SPI
interface follows the "Raiden" protocol, and flashrom and other tools
know how to recognize that, so we also do not want to change that.  This
leaves only the I2C interface.  I do not expect tools outside of
opentitantool to want to directly access the I2C interface, which means
that by making opentitantool able to recognize the CMSIS-DAP interface
as an alternative to the uServo-compatible I2C interface, it will be
able to speak to both old and new HyperDebug firmware.

BUG=b:299669075
TEST=tast run - gscdevboard.Ti50Tpm.i2c (with opentitantool changes)

Change-Id: I60c13562cf9b9b9991679eaa9fe40853dff882cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4852251
Reviewed-by: Jett Rink <jettrink@chromium.org>
Tested-by: Jes Klinke <jbk@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Jes Klinke <jbk@chromium.org>