lotus: Make sure switch_status is always accurate
Previously it would often get stuck in switch_status=true (open) even
if it's actually closed. The variable was only updated in the
`check_gpu_module hook`. We don't want to always run the hook when the
hatch switch is closed, only when it opens so that the power can be cut
off to the GPU.
But we still want to know the switch state, so set that manually without
hook.

Also remove the switch code in the `project_chassis_function` that's
called when the LID (not hatch) switch is opened. That was necessary
only in EVT.

Signed-off-by: Daniel Schaefer <dhs@frame.work>