i2chid fixup report descriptor for i2c int
When the system fast boots from hibernate, it will not
do a full descriptor read, and set the report id since
als only has a single endpoint.

Since we fire an interrupt, we have to set the
report id in this case, since the os does not init
the report id.

Such as when resuming from S4.

It seems that the OS no longer will respond to a
reset response (all 0x00) to reconfigure the als
report type.

Signed-off-by: Kieran Levin <ktl@frame.work>