timer: Prevent frequent error message
Calling usleep from an interrupt, or while interrupts are disabled,
results in a printing a "Sleeping not allowed" warning message. If
usleep is called too frequently this can fill the uart buffer and
result in lost messages. Reduce the frequency of this message to a
frequency of at most 20ms.

Ideally, calls to usleep would not occur if interrupts are disabled or
within an interrupt handler. This is difficult to achieve for low level
functions that may be called from either context, such as
flash_wait_ready from npcx/flash.c.

BUG=b:291141229
TEST=run_device_tests.py --board helipilot -t rollback_entropy
TEST=run_device_tests.py --board helipilot -t timer

Change-Id: I6509f0b9d606aad4f2ce4047602bdd720c650030
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4897166
Reviewed-by: David Cross <davidmcross@google.com>
Tested-by: Bobby Casey <bobbycasey@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Bobby Casey <bobbycasey@google.com>