docs: Use ChromiumOS SDK toolchain in library-compiler-flags.md
Add note about using ChromiumOS SDK toolchain, since it is too difficult
to use an external toolchain. This is because of the manual setup required
to user newlib and other compiler libraries, correctly.

We would be reinventing the wheel if we gave instructions on how to
build the toolchain from scratch.

BUG=b:301146101,b:306454550
TEST=Viewed in Gitiles

Change-Id: I72f67cdc88dee82a408880240fd9fa38838a4c10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4956465
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Bobby Casey <bobbycasey@google.com>
Commit-Queue: Craig Hesling <hesling@chromium.org>