fpmcu: Introduce BOARD_FINGERPRINT config
The config is used to indicate that this is the Fingerprint MCU. It's
selected by board config (e.g. BOARD_BLOONCHIPPER) and enabled in both
RO and RW, so it can be used to control 'fingerprint feature' bit.

BUG=b:299924239
TEST=zmake build bloonchipper

Change-Id: I5f56b14854b39f63a2e61869fdf660dcee7fa6bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4931236
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Patryk Duda <patrykd@google.com>
Tested-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>