zephyr/shim: prevent spurious failure of hook_call_deferred
Zephyr currently defines K_NO_WAIT to be the same as zero ticks, so
without this change it's possible to cause spurious failures in work
scheduling when:

1. A work item is started but preempted or the following steps are
   executed by its handler function.
2. The item is cancelled, putting it in the CANCELING state because its
   handler is running but has been preempted.
3. It is rescheduled again before the handler completes and the
   requested delay is zero to execute again as soon as possible.

When K_NO_WAIT is accidentally passed to k_work_reschedule in this
situation (easy to do because the value zero used for that symbol is
hidden in documentation and zero has a different meaning to the legacy
hook_call_deferred API than it does to Zephyr), it will return failure
rather than scheduling the work again because K_NO_WAIT has special
meaning.

Battery cutoff tests are modified because they previously depended on
pending_cutoff_deferred failing to schedule and leaving battery cutoff
IN_PROGRESS; they are fixed to ensure the AC state is initialized to a
known state for every test and to correctly check the cases which
actually do cause the battery to be marked as cut off, as well as
leaving the battery connected afterward.

BUG=b:305126895
TEST=twister;
     Newly-added test fails without also changing shim/src/hooks.c

Change-Id: I6bdef0f02c09a67e97b4ff9f6084bccbf2c3ca8c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4975896
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Peter Marheine <pmarheine@chromium.org>