fpsensor: Remove EC_CMD_FP_PASSTHRU host command handler
This command was used for prototyping in early FP development stages.
It's not available neither in ectool nor in libec, so there is no
easy way to use (and test) that. It also uses EC's SPI API which is not
compatible with Zephyr.

BUG=b:299924239
TEST=make BOARD=bloonchipper

Change-Id: I1da55e7cad7074ba2d465a08aeb887aef1b3da78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4987330
Commit-Queue: Patryk Duda <patrykd@google.com>
Tested-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Bobby Casey <bobbycasey@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>