chip/stm32: Properly declare clock and voltage ranges of STM32L5
So far, STM32L5 family has used the same clock-stm32l4.c as the L4
family.

Closely reading the section 6.3.3 of reference manual RM0438 reveals
though that the flash memory in L5 family has a table of required wait
states as function of clock frequency which differs from the one of L4.

Furthermore, it turns out that the default "Voltage scaling range" of
the MCU core is "range 2" (the lowest performance range) on the L5
series, whereas it is "range 1" (the highest performance range) on L4.
This necessitates some differences, so with this CL I first make a copy
of clock-stm32l4.c to become clock-stm32l5.c (patchset 1), and then I
have edited the clock setup function to match L5.

(There is a separate issue with the original L4 code, it seems that it
writes to register STM32_RCC_CR when the intention clearly was to write
to STM32_PWR_CR1, to switch voltage range.  This bug is mitigated by the
fact that L4 starts up in the high performance range 1, meaning that no
code is needed to switch to range 1.  On the contrary, the low power
branch could possibly switch to range 2, if we wanted to improve power
usage on L4.)

Change-Id: I275a4697f83f26172cab1eb85abc23cb6ce8ee8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4990075
Commit-Queue: Jes Klinke <jbk@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Jes Klinke <jbk@chromium.org>