tests: exercise smart battery errors with cutoff
This adds a test to exercise the error paths in smart battery accessors
which returns an error if the battery is cut off, which was covered by
happenstance prior to crrev.com/c/4975896 because the battery got stuck
pending cutoff.

The smart battery tests need to be moved to a new suite because other
tests in the drivers.default suite depend on having real battery cutoff
functions and the new test case needs to mock it out.

Uses of EC_RES_ACCESS_DENIED are changed to EC_ERROR_ACCESS_DENIED to
make the battery functions consistent in their error types; EC_RES_*
values are part of the host command API whereas EC_ERROR_* is the
EC-internal error type.

BUG=b:308185587
TEST=twister

Change-Id: I37c5ca1833e19b06253a021866518496bd8cd192
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4990651
Commit-Queue: Peter Marheine <pmarheine@chromium.org>
Tested-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>