zephyr: Support deduplicating RW FWID
The EC version is placed in current_image_data.version[32] for each
RO/RW firmware blob.

In CrOS-EC, the FWID fmap section points to the above version[32].
So other tools can easily read the version string from the EC image.

But in Zephyr-EC, a separated FWID fmap section is created which
is put at the end of the RO_FW/RW_FW sections. It leads problems:
 * Duplication of the version string.
 * The size of the RW_FW section (the size to update) is the same
   size as the EC_RW section (the available flash size). It adds
   time penalty to the boot time (calculating the hash) and the
   EC update time.

This CL adds the support of the CrOS-EC way for RW FWID. For
projects want to use it, they can include the `integrated_fwid.dtsi`
overlay. The existing projects, which don't include the overlay,
have no change.

The RO FWID keeps the same. As the RO image may add a chip-specific
header to the beginning so the loader understands how to load the RO
image from the flash. This header shifts the elf symbols. A solution
needs to be studied.

A tast test to check the fmap layout, like inclusion, overlap, is in
tast-tests/cros/remote/bundles/cros/firmware/fmap.go

BUG=b:198475757
TEST=With the following changes, checked the RW_FWID fmap sections
point to current_image_data.version[32].

Change-Id: I1e7de14c1d71300b39933cba0a23a6b80972ae8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4961894
Tested-by: Wai-Hong Tam <waihong@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>