zephyr: Use the effective EC RW firmware size to compute hash
Briefly give a background of 2 concepts of sizes:
 * EC-RW: The allocated size, that is the amount of flash space
          allocated to store the RW firmware.
 * RW-FW: The effective size, that stores the actual firmware
          which is updated during EC software sync.

The EC EFS2 uses the RW's current_image_data.size to compute the
hash. This size variable should be the same as the RW-FW fmap so
it matches with what is updated.

However, this size variable is currently assigned to the EC-RW
fmap. It works fine because the EC-RW and the RW-FW have the same
size. We put the RW-FWID at the end of the EC-RW and the RW-FWID
is contained within the RW-FW. But no longer true if we make the
RW-FWID pointing to current_image_data.version.

We can't simply change the size variable to the RW-FW fmap because
its size is not explicitly assigned in the devicetree. Its size
is determined after binman packs the entire image.

This CL introduces a new Kconfig which indicates the RW-FWID overlaps
with the firmware blob. If the Kconfig is enabled, set the size
variable to the linker symbol _flash_used or __rom_region_size.

Add the missing properties in the "cros-ec,flash-layout" binding as
the RW-FWID node now needs to be parsed.

BUG=b:198475757
TEST=With and without the following changes to enable FWID overlap,
decoded the EC binary and checked the current_image_data.size has
the same value as RW-FW fmap.

Verified on a ARM device and a RISC-V device.

Change-Id: I15efabc86351999f58465a521619229d1e9916d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5004987
Commit-Queue: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>