chip/stm32: Remove confusing code with no effect
The STM32L4/L5 declares STM32_MSI_CLOCK to default to 4MHz, and use that
value for computing e.g. UART baud divisors.

It is confusing then, that the code to switch to MSI clock source
explicitly sets a register to STM32_RCC_ICSCR_MSIRANGE_1MHZ, which would
seem to suggest that the MSI oscillator would be programmed to generate
1MHz not 4MHz.  It turns out, though, that since the code does not also
set the MSIRGSEL bit, the MSI oscillator ignores the 1MHz setting and
runs on its default.

This CL keeps the 4MHz constant, but removes the code that looks as if
the frequency is set at 1MHz.

(Also, remove some redundant declarations in registers, and add
meaningful labels to others.)

BUG=b:309245084
TEST=Observe HyperDebug still working on MSI clock

Change-Id: Ic140b8e9cea9372e3c9a17021fdeea4c29519c7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5005632
Commit-Queue: Jes Klinke <jbk@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Jes Klinke <jbk@chromium.org>
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>