extra: i2c_pseudo: Documentation: `make` instead of `cc`
In the sample i2c-adapter-example usage this replaces:
$ cc -o i2c-adapter-example.c i2c-adapter-example

With:
$ make i2c-adapter-example

This is the correct way to build it using the local source tree
i2c-pseudo.h UAPI header, instead of depending on there being a
preinstalled system copy.

The make command is functional and correct with both the
standalone Makefile introduced in CL:5031761, and with the
Makefile in the upcoming patch for upstream review.

BUG=b:275615681
TEST=I tested the changed lines in the instructions:

$ cd samples/i2c-pseudo
$ make i2c-adapter-example

Change-Id: I3f7a51c3d9a062ea7d03d77cd56caca703b516ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5035544
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Matthew Blecker <matthewb@chromium.org>
Tested-by: Matthew Blecker <matthewb@chromium.org>
Reviewed-by: Brian Nemec <bnemec@google.com>