zephyr: keyboard: use the correct kso02 logical level
The line is normally active low to select the column, but being inverted
make it active high (i.e. the column is selected when the signal is at
the high level), so the right way of declaring it is flagging as
GPIO_ACTIVE_HIGH.

Change all current board to flag the kso02 column as GPIO_ACTIVE_HIGH
and change the npcx and xec driver to use active and inactive correctly.

The ITE driver is not affected as that uses a special Kconfig option.

BUG=b:274443029
TEST=zmake buildall
TEST=test on rex
LOW_COVERAGE_REASON=b:256669239 these seems to be out of scope, working
on replacing them with the upstream drivers anyway

Change-Id: Id9a6162f2881a20766727dde4652e885257a15bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5049862
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>