ap_pwrseq: enable driver model by default
The driver variant of ap_pwrseq is meant to be the long-term solution
for power sequencing by improving on a number of API aspects. However,
it's unlikely that existing platforms will migrate to it due to the risk
involved in changing power sequencing implementations. To encourage new
platforms to use the correct implementation, change the driver model to
be the default and explicitly disable it for existing platforms that use
ap_pwrseq.

BUG=b:217952699
TEST=zmake compare-builds -a

Change-Id: I76208bae9bc1bc0f9ab99db5314fb64145811871
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5069581
Reviewed-by: Bernardo Perez Priego <bernardo.perez.priego@intel.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Peter Marheine <pmarheine@chromium.org>
Tested-by: Peter Marheine <pmarheine@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>