binman: Add the padding of EC_RW region
The EC_RW region contains the RW_FW region. The RW_FW used to be
the same size as the EC_RW as the RW_FWID is set at the end of the
EC_RW. So no padding is needed for the EC_RW.

Once we move the RW_FWID to an overlapped region pointing to the
current_image_data, the RW_FW is smaller than the EC_RW. The gap
shows the padding with 0. This CL fills it with 0xff.

BUG=b:198475757
TEST=Built the Rex image and check the padding:
Before:
$ hexdump ec.bin | tail -n 6
00c1650 1388 0000 e015 e015 0000 0000 0000 0000
00c1660 0000 0000 0000 0000 0000 0000 0000 0000
*
00d0000 ffff ffff ffff ffff ffff ffff ffff ffff
*
0100000

After:
$ hexdump ec.bin | tail -n 6
00c1630 1771 0000 f879 1007 1771 0000 a709 1008
00c1640 1388 0000 a499 1007 1388 0000 74d9 1007
00c1650 1388 0000 e015 e015 ffff ffff ffff ffff
00c1660 ffff ffff ffff ffff ffff ffff ffff ffff
*
0100000

Change-Id: Ie32201fb877ffe3bc88c63902185766ada2fb449
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5119136
Commit-Queue: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>